@@ -19,10 +19,7 @@ use HeadlessChromium\BrowserFactory;
  */
 class KeyboardApiTest extends BaseTestCase
 {
-    /**
-     * @var Browser\ProcessAwareBrowser
-     */
-    public static $browser;
+    public static Browser\ProcessAwareBrowser $browser;
 
     public static function setUpBeforeClass(): void
     {
